readme

+++ engine/core/animation-plane.hpp
@@ -49,7 +49,7 @@ namespace hid
 
         AnimationPlane(
             const hid::Material &material,
-            const float &animationFrame,
+            const uint32_t &animationFrame,
             const glm::vec3 &position = glm::vec3{0.0f, 0.0f, 0.0f},
             const glm::vec3 &scale = glm::vec3{1.0f, 1.0f, 1.0f},
             const glm::vec3 &rotationAxis = glm::vec3{0.0f, 1.0f, 0.0f},
